# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HealthAI Coach API",
    "description": "\n    API REST du backend métier HealthAI Coach.\n\n    Gère les données de nutrition, fitness et utilisateurs\n    pour la plateforme de coaching santé personnalisé.\n    ",
    "contact": {
      "name": "Équipe HealthAI Coach"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/coach/advice": {
      "post": {
        "tags": [
          "Coach IA"
        ],
        "summary": "Conseil personnalisé IA",
        "description": "Génère un conseil de coaching santé personnalisé via OpenAI GPT à partir du profil et de la dernière mesure biométrique de l'utilisateur.",
        "operationId": "get_ai_advice_coach_advice_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoachRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/coach/analyze-photo": {
      "post": {
        "tags": [
          "Coach IA"
        ],
        "summary": "Analyse photo de repas (Vision IA)",
        "description": "Soumet une photo de repas encodée en base64 à GPT-4o Vision. Retourne les aliments détectés, une estimation des macronutriments et un conseil nutritionnel adapté à l'objectif de l'utilisateur.",
        "operationId": "analyze_photo_coach_analyze_photo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/coach/workout-plan": {
      "post": {
        "tags": [
          "Coach IA"
        ],
        "summary": "Programme d'entraînement personnalisé",
        "description": "Génère un programme d'entraînement hebdomadaire via GPT-4o-mini, adapté à l'objectif de l'utilisateur, au matériel disponible et au nombre de séances souhaité.",
        "operationId": "get_workout_plan_coach_workout_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/coach/meal-plan": {
      "post": {
        "tags": [
          "Coach IA"
        ],
        "summary": "Plan repas personnalisé avec contraintes",
        "description": "Génère un plan repas hebdomadaire via GPT-4o-mini, adapté à l'objectif de l'utilisateur, à son budget et à ses allergies ou intolérances alimentaires.",
        "operationId": "get_meal_plan_coach_meal_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/coach/biometric-trend": {
      "post": {
        "tags": [
          "Coach IA"
        ],
        "summary": "Analyse des tendances biométriques (30 jours)",
        "description": "Récupère les 30 derniers jours de métriques biométriques de l'utilisateur et demande à GPT d'analyser les tendances (poids, sommeil, FC repos) avec des conseils adaptés à l'objectif.",
        "operationId": "get_biometric_trend_coach_biometric_trend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoachRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoachRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Identifiant de l'utilisateur en base"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "CoachRequest",
        "example": {
          "user_id": 1
        }
      },
      "CoachResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "advice": {
            "type": "string",
            "title": "Advice"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "advice"
        ],
        "title": "CoachResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealPlanRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Identifiant de l'utilisateur"
          },
          "budget_euros": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 10.0,
            "title": "Budget Euros",
            "description": "Budget hebdomadaire en euros (10–500)",
            "default": 50.0
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies",
            "description": "Liste des allergies ou intolérances (ex: ['gluten', 'lactose'])"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "MealPlanRequest",
        "example": {
          "allergies": [
            "gluten",
            "lactose"
          ],
          "budget_euros": 50.0,
          "user_id": 1
        }
      },
      "MealPlanResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "plan"
        ],
        "title": "MealPlanResponse"
      },
      "PhotoRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Identifiant de l'utilisateur"
          },
          "image_base64": {
            "type": "string",
            "title": "Image Base64",
            "description": "Image encodée en base64 (JPEG ou PNG)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "image_base64"
        ],
        "title": "PhotoRequest",
        "example": {
          "image_base64": "<base64>",
          "user_id": 1
        }
      },
      "PhotoResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "foods_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Foods Detected"
          },
          "macros": {
            "$ref": "#/components/schemas/Macros"
          },
          "advice": {
            "type": "string",
            "title": "Advice"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "foods_detected",
          "macros",
          "advice"
        ],
        "title": "PhotoResponse"
      },
...
      "TrendResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "analysis": {
            "type": "string",
            "title": "Analysis"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "analysis"
        ],
        "title": "TrendResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkoutRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Identifiant de l'utilisateur"
          },
          "equipment": {
            "type": "string",
            "title": "Equipment",
            "description": "Matériel disponible : none | dumbbell | barbell | machine | resistance | full",
            "default": "none"
          },
          "days_per_week": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Days Per Week",
            "description": "Nombre de séances par semaine (1–7)",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "WorkoutRequest",
        "example": {
          "days_per_week": 3,
          "equipment": "dumbbell",
          "user_id": 1
        }
      },
      "WorkoutResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "plan"
        ],
        "title": "WorkoutResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}